requirements.txt
fastapi
uvicorn
pydantic
requests
groq
gitpython
python-dotenv
